requirements.txt
dnspython
python-whois
tqdm
pandas